Move allow/deny rules check to AuthenticationPlugin
Issue #11731
